Lots of refactoring and documentation added. Adjusted ButtonInputParser so it should work better with modifying actions. Haven't really tested it, but it seems to work perfectly at least for my current configuration.

+++ Mouse Fix Helper/Remap/Input/ButtonInputReceiver.m
@@ -95,7 +95,7 @@ CGEventRef handleInput_CG(CGEventTapProxy proxy, CGEventType type, CGEventRef ev
     _buttonEventInputSourceIsDeviceOfInterest = false;
     
     if (b) {
-        int buttonNumber = (int) CGEventGetIntegerValueField(event, kCGMouseEventButtonNumber) + 1;
+        int64_t buttonNumber = CGEventGetIntegerValueField(event, kCGMouseEventButtonNumber) + 1;
         
         long long pr = CGEventGetIntegerValueField(event, kCGMouseEventPressure);
         MFButtonInputType type = pr == 0 ? kMFButtonInputTypeButtonUp : kMFButtonInputTypeButtonDown;
